Refactor serialization logic for clarity
This commit introduces the `MetadataObjectFormatter` class, implementing the `IMessagePackFormatter<object?>` interface to centralize serialization and deserialization logic for a wide range of types, including special cases like `Enum32Substitution`. It significantly refactors the `MetadataDictionaryFormatter`, extracting redundant logic into the new class to adhere to the single responsibility principle. Unused usings and obsolete serialization/deserialization code have been removed from `MetadataDictionaryFormatter.cs` for better readability. Additionally, `MetadataObjectFormatter.Instance` is now utilized in both `MetadataDictionaryFormatter` and `MessagePackSerializerContext` for serialization tasks, ensuring a unified approach to object formatting. The handling of `null` values during serialization and deserialization has been improved. The `MessagePackSerializerContext` class has been updated to include more types in its `DedupingTypes` set and to incorporate `MetadataObjectFormatter.Instance` in its resolver setup, optimizing serialization efficiency and consistency across the application.
